[2~4단계 - 방탈출 결제 / 배포] 시소(이경민) 미션 제출합니다. (#325)
* refactor: 메시지 비교 NPE 방지

* refactor: 미사용 lombok annotation 제거

* feat: 에러 로그 통일

* refactor: 미사용 주석 제거

* feat: Payment Entity 추가

* docs: 변경된 기능 목록 재정의

* feat: 결제 정보 저장 및 내 예약 정보와 함께 결제 정보 조회 기능 추가

* refactor: 미사용 주석 제거

* feat: 내 예약 대기 삭제 기능 추가

* feat: 엔티티 필드 제약 조건 추가

* feat: 로그 추가 및 파일 저장 기능 추가

* feat: 로그 추가 및 파일 저장 기능 추가

* feat: Swagger API Docs 추가

* test: 추가 기능에 대한 테스트 추가

* fix: level이 ERROR인 log만 파일에 저장하도록 수정

* refactor: 변수명 오타 수정

* test: DisplayName 추가 및 예외 메시지 검증 테스트 추가

* feat: 토스 결제 API 예외 처리 분기 추가 및 로그 추가

* feat: 예외 메시지에 ID 포함

* docs: API 문서 설명 추가

* feat: 배포 스크립트 작성

* refactor: 예약 상태 저장 CODE 변경

+++ src/main/java/roomescape/reservation/service/usecase/ReservationCommandUseCase.java
@@ -1,16 +1,21 @@
 package roomescape.reservation.service.usecase;
 
+import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.common.exception.AlreadyExistException;
+import roomescape.common.exception.NotFoundException;
 import roomescape.member.domain.Member;
 import roomescape.member.service.usecase.MemberQueryUseCase;
+import roomescape.payment.domain.Payment;
+import roomescape.payment.service.usecase.PaymentQueryUseCase;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationDate;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.service.converter.ReservationConverter;
 import roomescape.reservation.service.dto.CreateReservationServiceRequest;
+import roomescape.reservation.service.dto.CreateReservationWithPaymentServiceRequest;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.service.usecase.ThemeQueryUseCase;
 import roomescape.time.domain.ReservationTime;
@@ -26,28 +31,65 @@ public class ReservationCommandUseCase {
     private final ReservationTimeQueryUseCase reservationTimeQueryUseCase;
     private final ThemeQueryUseCase themeQueryUseCase;
     private final MemberQueryUseCase memberQueryUseCase;
+    private final PaymentQueryUseCase paymentQueryUseCase;
 
-    public Reservation create(final CreateReservationServiceRequest createReservationServiceRequest) {
-        if (reservationQueryUseCase.existsByParams(
-                ReservationDate.from(createReservationServiceRequest.date()),
-                createReservationServiceRequest.timeId(),
-                createReservationServiceRequest.themeId())) {
-
-            throw new AlreadyExistException("추가하려는 예약이 이미 존재합니다.");
-        }
+    public Reservation create(final CreateReservationServiceRequest request) {
+        validateExistReservation(request.date(), request.timeId(), request.themeId());
 
         final ReservationTime reservationTime = reservationTimeQueryUseCase.get(
-                createReservationServiceRequest.timeId());
+                request.timeId());
 
-        final Theme theme = themeQueryUseCase.get(createReservationServiceRequest.themeId());
+        final Theme theme = themeQueryUseCase.get(request.themeId());
 
-        final Member member = memberQueryUseCase.get(createReservationServiceRequest.memberId());
+        final Member member = memberQueryUseCase.get(request.memberId());
 
         return reservationRepository.save(
-                ReservationConverter.toDomain(createReservationServiceRequest, member, reservationTime, theme));
+                ReservationConverter.toDomain(
+                        request,
+                        member,
+                        reservationTime,
+                        theme
+                )
+        );
+    }
+
+    public Reservation create(final CreateReservationWithPaymentServiceRequest request) {
+        validateExistReservation(request.date(), request.timeId(), request.themeId());
+
+        final ReservationTime reservationTime = reservationTimeQueryUseCase.get(
+                request.timeId());
+
+        final Theme theme = themeQueryUseCase.get(request.themeId());
+
+        final Member member = memberQueryUseCase.get(request.memberId());
+
+        final Payment payment = paymentQueryUseCase.get(request.paymentId());
+
+        return reservationRepository.save(
+                ReservationConverter.toDomain(
+                        request,
+                        member,
+                        reservationTime,
+                        theme,
+                        payment
+                )
+        );
     }
 
     public void delete(final Long id) {
+        if (!reservationRepository.existsById(id)) {
+            throw new NotFoundException("id: %d에 해당하는 예약을 찾을 수 없습니다.", id);
+        }
         reservationRepository.deleteById(id);
+    }
+
+    private void validateExistReservation(final LocalDate date, final Long timeId, final Long themeId) {
+        if (reservationQueryUseCase.existsByParams(
+                ReservationDate.from(date),
+                timeId,
+                themeId)
+        ) {
+            throw new AlreadyExistException("추가하려는 예약이 이미 존재합니다.");
+        }
     }
 }
